make sure the latest version is running as server

+++ bin/test.wdio.sh
@@ -9,8 +9,16 @@ npx webpack --config ./specs.wdio/testapp/webpack.config.js
 
 bin/testserver.sh&
 SERVER_PID=$!
-echo "Node server is running on with PID ${SERVER_PID}"
+
 sleep 2
+
+if ! ps -p $SERVER_PID; then
+  echo "Node test server is NOT running!!! Maybe the server is running already and blocking the port. Check for 'ps aux | grep testserver' and terminate it."
+  exit 1
+fi
+
+echo "Node server is running on with PID ${SERVER_PID}"
+
 npx wdio run ./wdio.conf.ts
 succ=$?
 kill $SERVER_PID
